Move test helpers into own file.
